requirements.txt

tox==2.7.0
Flask==0.12.2
flask-cors==3.0.2
flask-pytest == 0.0.5
pytest == 5.2
pylint == 2.5.3
pylint-flask == 0.6
pylint_flask_sqlalchemy == 0.2.0
marshmallow-enum == 1.5.1
py-postgresql == 1.2.1
pytest-cov >= 2.0
pytest-filter-subpackage>=0.1
Flask-Migrate == 2.5.3
flask-marshmallow == 0.13.0
psycopg2 == 2.8.5